src/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Jumageldiyew Batyr <jumageldiyewb@list.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.74.0"
uvicorn = "^0.17.5"
gunicorn = "^20.1.0"
Jinja2 = "^3.0.3"
emails = "^0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
email-validator = "^1.1.3"
python-multipart = "^0.0.5"
python-magic = "^0.4.25"
orjson = "^3.6.7"
Babel = "^2.9.1"
starlette-i18n = "^1.0.0"
python-dotenv = "^0.19.2"
WTForms = "^3.0.1"
python-slugify = "^6.0.1"
Markdown = "^3.3.6"
itsdangerous = "^2.1.0"
humanize = "^4.0.0"
SQLAlchemy = "^1.4.31"
alembic = "^1.7.6"
WTForms-Alchemy = "^0.18.0"
SQLAlchemy-Utils = "^0.38.2"
psycopg2-binary = "^2.9.3"
phonenumbers = "^8.12.43"
asgi-sitemaps = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
